refactor(reporter): extract validator cap ante logic
Move active-set projection and validator power share checks into a dedicated
ante file and simplify unjail tracking for M2 isolation.
