+++ x/reporter/ante/reporter_power_cap_test.go
@@ -212,9 +212,7 @@ func TestReporterPowerCapDelegateBySelector(t *testing.T) {
 			decorator := NewTrackStakeChangesDecorator(k, sk)
 			// the bonded validator holds 1000 of 100 total bonded; this test targets
 			// the reporter cap, so disable the validator cap.
-			params := types.DefaultParams()
-			params.MaxValidatorPowerShare = math.LegacyOneDec()
-			require.NoError(t, k.Params.Set(ctx, params))
+			disableValidatorPowerCap(t, k, ctx)
 
 			valAddr := sdk.ValAddress(sample.AccAddressBytes())
 			val := validator(valAddr, stakingtypes.Bonded, math.NewInt(1000))
@@ -252,9 +250,7 @@ func TestReporterPowerCapSelectPlusDelegate(t *testing.T) {
 	decorator := NewTrackStakeChangesDecorator(k, sk)
 	// the bonded validator holds 1000 of 100 total bonded; this test targets the
 	// reporter cap, so disable the validator cap.
-	params := types.DefaultParams()
-	params.MaxValidatorPowerShare = math.LegacyOneDec()
-	require.NoError(t, k.Params.Set(ctx, params))
+	disableValidatorPowerCap(t, k, ctx)
 
 	valAddr := sdk.ValAddress(sample.AccAddressBytes())
 	val := validator(valAddr, stakingtypes.Bonded, math.NewInt(1000))
